Add files via upload
Identify Cat And Dog Owners Find the users who are cat and/or dog owners.

Build And Evaluate Classifiers Build classifiers for the cat and dog owners and measure the performance of the classifiers.

Classify All The Users Apply the cat/dog classifiers to all the users in the dataset. Estimate the fraction of all users who are cat/dog owners.

Extract Insights About Cat And Dog Owners Find topics important to cat and dog owners.

Identify Creators With Cat And Dog Owners In The Audience Find creators with the most cat and/or dog owners. Find creators with the highest statistically significant percentages of cat and/or dog owners.